[Feat] 레시피 목록 조회 추가

+++ src/main/java/store/myproject/onlineshop/controller/LikeController.java
@@ -28,8 +28,8 @@ public class LikeController {
 
     @Operation(summary = "해당 레세피 좋아요 개수 조회")
     @GetMapping("/{id}/likes")
-    public Response<Integer> countLike(@PathVariable Long id) {
-        Integer likeCnt = recipeService.getLikeCount(id);
+    public Response<Long> countLike(@PathVariable Long id) {
+        Long likeCnt = recipeService.getLikeCount(id);
         return Response.success(likeCnt);
     }
